feat: return error code if failed

+++ examples/queue_multiple_transactions_and_poll_all_results/queue_multiple_transactions_and_poll_all_results_master/queue_multiple_transactions_and_poll_all_results_master.ino
@@ -61,7 +61,7 @@ void loop()
         Serial.println("all queued transactions completed. start verifying received data from slave");
 
         // get received bytes for all transactions
-        const std::vector<size_t> received_bytes = master.numBytesReceivedAll();
+        const std::vector<int64_t> received_bytes = master.numBytesReceivedAll();
 
         // verify and dump difference with received data
         // NOTE: we need only 2nd results (received_bytes[1])
